+++ R/emojis.R
@@ -1,0 +1,8 @@
+#' Emojis
+#'
+#' Useful emojis
+#'
+#' * General: "check", "star", warning", "stop", "x", "no", "thumbs_up", "thumbs_down", "info", "question"
+#' * Traffic Lights: "tl_green", "tl_yellow", "tl_red", "tl_info", "tl_na", "tl_fail"
+#' * Hearts: "red", "orange", "yellow", "green", "blue", "purple", "brown", "black", "white", "pink"
+"emojis"