+++ R/utils.R
@@ -4,6 +4,7 @@
 #'
 #' @param x,y If `x` is NULL, will return `y`; otherwise returns `x`.
 #' @export
+#' @keywords internal
 #' @name op-null-default
 #' @examples
 #' 1 %||% 2