+++ tests/testthat/test-07-llm.R
@@ -178,3 +178,24 @@ test_that("llm_model_list", {
   expect_equal(names(models), c("id", "owned_by", "created", "context_window"))
   expect_true(llm_model() %in% models$id)
 })
+
+test_that("json_expand", {
+  table <- data.frame(
+    id = 1:5,
+    answer = c(
+      '{"number": "1", "letter": "A", "bool": true}',
+      '{"number": "2", "letter": "B", "bool": "FALSE"}',
+      '{"number": "3", "letter": "", "bool": null}',
+      'oh no, the LLM misunderstood',
+      '{"number": "5", "letter": ["E", "F"], "bool": false}'
+    )
+  )
+
+  expanded <- json_expand(table)
+  expect_equal(names(expanded), c("id", "answer", "number", "letter", "bool", "error"))
+  expect_equal(typeof(expanded$number), "integer")
+  expect_equal(typeof(expanded$letter), "character")
+  expect_equal(typeof(expanded$bool), "logical")
+  expect_equal(expanded$letter, c("A", "B", "", NA, "E; F"))
+  expect_equal(expanded$bool, c(TRUE, FALSE, NA, NA, FALSE))
+})